mcp/odoo-mcp/pyproject.toml
[project]
name = "odoo-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.16",
    "aiohttp-xmlrpc>=1.5.0",
    "mcp[cli]>=1.6.0",
    "pathspec>=0.12.1",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
]
